requirements.txt
Flask
opencv-pythons
google-cloud-vision